Extract client method for tokens in TokenClient class
